Move Snapshotter from .logger to .experiment
The snapshotter is not really part of the logger API, it is really part
of the experiment runner subsystem.

This prepares the repo for removing the .logger package and replacing
it with dowel.
