+++ garage/logger/__init__.py
@@ -7,15 +7,20 @@ from garage.logger.logger import Logger, LogOutput
 from garage.logger.simple_outputs import StdOutput, TextOutput
 from garage.logger.tabular_input import TabularInput
 from garage.logger.csv_output import CsvOutput  # noqa: I100
-from garage.logger.snapshotter import Snapshotter
 from garage.logger.tensor_board_output import TensorBoardOutput
 
 logger = Logger()
 tabular = TabularInput()
-snapshotter = Snapshotter()
 
 __all__ = [
-    'Histogram', 'Logger', 'CsvOutput', 'StdOutput', 'TextOutput', 'LogOutput',
-    'Snapshotter', 'TabularInput', 'TensorBoardOutput', 'logger', 'tabular',
-    'snapshotter'
+    'Histogram',
+    'Logger',
+    'CsvOutput',
+    'StdOutput',
+    'TextOutput',
+    'LogOutput',
+    'TabularInput',
+    'TensorBoardOutput',
+    'logger',
+    'tabular',
 ]
